test(op-dispatch): merge tcgen05.cp test files into test_tcgen05_cp.py
Fold test_smem_tmem.py (legacy 32x128b.warpx4 tests) into the generic-planner
test file as a dedicated section, renamed test_tcgen05_cp_shapes.py ->
test_tcgen05_cp.py, and updated filename references in the planner docstring
and the rst doc. No test semantics changed; 118 tests pass on B200.
